requirements.txt
aiogram==2.18
albumentations==0.5.2
dlib==19.23.0
numpy<2
opencv-python
pandas
Pillow==9.1.0
plotly==5.6.0
scipy==1.7.3
torch
torchvision
tensorflow
